requirements.txt
Django>=5.1.5
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==10.2.0 